Fix flash assignments before render
